+++ src/main/java/Fo/Suzip/apiPayload/exception/handler/MyAuthenticationSuccessHandler.java
@@ -1,14 +1,13 @@
 package Fo.Suzip.apiPayload.exception.handler;
 
 import Fo.Suzip.jwt.JwtUtil;
-import Fo.Suzip.web.dto.CustomOAuth2User;
-import Fo.Suzip.web.dto.GeneratedToken;
+import Fo.Suzip.web.dto.authDTO.CustomOAuth2User;
+import Fo.Suzip.web.dto.authDTO.GeneratedToken;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
