+++ src/main/java/Fo/Suzip/jwt/JwtAuthFilter.java
@@ -1,11 +1,8 @@
 package Fo.Suzip.jwt;
 
 import Fo.Suzip.converter.MemberConverter;
-import Fo.Suzip.domain.Member;
-import Fo.Suzip.repository.MemberRepository;
-import Fo.Suzip.service.MemberService;
-import Fo.Suzip.web.dto.CustomOAuth2User;
-import Fo.Suzip.web.dto.SecurityUserDto;
+import Fo.Suzip.web.dto.authDTO.CustomOAuth2User;
+import Fo.Suzip.web.dto.authDTO.SecurityUserDto;
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -13,18 +10,13 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
-import java.util.List;
 
 @RequiredArgsConstructor
 //@Slf4j
