+++ src/main/java/Fo/Suzip/service/DiaryService/DiaryService.java
@@ -1,17 +1,12 @@
 package Fo.Suzip.service.DiaryService;
 
 import Fo.Suzip.domain.Diary;
-import Fo.Suzip.repository.DiaryRepository;
-import Fo.Suzip.web.dto.DiaryDTO;
+import Fo.Suzip.web.dto.diaryDTO.DiaryDTO;
 import java.util.List;
 
-import Fo.Suzip.web.dto.DiaryRequestDTO;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import Fo.Suzip.web.dto.diaryDTO.DiaryRequestDTO;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.*;
 
 @Service
 @Transactional(readOnly = true) // 읽기 전용 트랜잭션 설정
