+++ src/main/java/Fo/Suzip/service/OAuth2Service/CustomOAuth2UserService.java
@@ -3,11 +3,9 @@ package Fo.Suzip.service.OAuth2Service;
 import Fo.Suzip.converter.MemberConverter;
 import Fo.Suzip.domain.Member;
 import Fo.Suzip.repository.MemberRepository;
-import Fo.Suzip.service.MemberService;
-import Fo.Suzip.web.dto.CustomOAuth2User;
-import Fo.Suzip.web.dto.SecurityUserDto;
+import Fo.Suzip.web.dto.authDTO.CustomOAuth2User;
+import Fo.Suzip.web.dto.authDTO.SecurityUserDto;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
