+++ src/main/java/Fo/Suzip/web/controller/ContentController.java
@@ -3,9 +3,8 @@ package Fo.Suzip.web.controller;
 import Fo.Suzip.apiPayload.ApiResponse;
 import Fo.Suzip.converter.ContentConverter;
 import Fo.Suzip.domain.contentItem.Book;
-import Fo.Suzip.domain.contentItem.ContentItem;
 import Fo.Suzip.service.contentService.ContentQueryService;
-import Fo.Suzip.web.dto.ContentResponseDTO;
+import Fo.Suzip.web.dto.contentDTO.ContentResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
