+++ src/main/java/Fo/Suzip/web/controller/MemberController.java
@@ -4,8 +4,8 @@ import Fo.Suzip.apiPayload.ApiResponse;
 import Fo.Suzip.converter.MemberConverter;
 import Fo.Suzip.domain.Member;
 import Fo.Suzip.service.MemberService;
-import Fo.Suzip.web.dto.MemberRequestDTO;
-import Fo.Suzip.web.dto.MemberResponseDTO;
+import Fo.Suzip.web.dto.memberDTO.MemberRequestDTO;
+import Fo.Suzip.web.dto.memberDTO.MemberResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
