--- src/main/java/Fo/Suzip/web/dto/ContentRequestDTO.java
@@ -1,4 +1,0 @@
-package Fo.Suzip.web.dto;
-
-public class ContentRequestDTO {
-}
