+++ src/main/java/Fo/Suzip/web/dto/authDTO/GeneratedToken.java
@@ -1,4 +1,4 @@
-package Fo.Suzip.web.dto;
+package Fo.Suzip.web.dto.authDTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
