+++ src/main/java/Fo/Suzip/web/dto/contentDTO/ContentRequestDTO.java
@@ -1,0 +1,4 @@
+package Fo.Suzip.web.dto.contentDTO;
+
+public class ContentRequestDTO {
+}
