+++ src/main/java/Fo/Suzip/web/dto/contentDTO/ContentResponseDTO.java
@@ -1,4 +1,4 @@
-package Fo.Suzip.web.dto;
+package Fo.Suzip.web.dto.contentDTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
