+++ src/main/resources/application-local.yml
@@ -20,7 +20,7 @@ spring:
         format_sql: true
         use_sql_comments: true
         hbm2ddl:
-          auto: create
+          auto: update
         default_batch_fetch_size: 1000
   security:
     user:
