+++ src/main/resources/application.yml
@@ -1,4 +1,4 @@
 spring:
   profiles:
     include: dev, local
-    active: local
+    active: dev
